+++ sdk/core/azure-core/CHANGELOG.md
@@ -13,7 +13,7 @@
 - Added `GetPort()` to `Url`.
 - Added `TransportPolicyOptions`.
 - Added `TelemetryPolicyOptions`.
-- Added `DateFormat::ToISO8601String(TimeFractionFormat)`.
+- Added `DateFormat::ToIso8601String(TimeFractionFormat)`.
 
 ### Other changes and Improvements